move warning component under src/app/shared/warning-message
