+++ templates/Web/Pages/Angular.MasterDetail/src/app/app-shell/AngularMasterDetail/AngularMasterDetail.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { MasterDetailComponent } from './master-detail.component';
 import { MasterDetailSideBarTabComponent } from './MasterDetailSideBarTab/master-detail-SideBarTab';
-import { WarningMessageModule } from 'src/app/sharedComponents/warning-message.module';
+import { WarningMessageModule } from 'src/app/shared/warning-message/warning-message.module';
 import { MasterDetailPageComponent } from './MasterDetailPage/master-detail-page.component';
 
 @NgModule({
