+++ templates/Web/Projects/AngularDefault/README_postaction.md
@@ -10,9 +10,9 @@ In the root directory of the project...
 //{[{
 ### Adding a New Page
 
-1. Create a folder in `/src/app/pages` with your angular modules.
+1. Create a folder in `/src/app/app-shell` with your angular modules.
 2. Add a child route for your page to `/src/app/*.module.ts`.
-3. Add a button to the navigation bar in `/src/components/NavBar/navbar.component.html`.
+3. Add a button to the navigation bar in `/src/app/app-shell/NavBar/navbar.component.html`.
 //}]}
 
 ### Deployment
