+++ templates/Web/_composition/AngularJS/Page.Angular.AddWarnings/src/app/shared/warning-message/warning-message.component.ts
